Add CSRF protection
Wrap the routerHandler with gorilla/csrf to enable CSRF protection
on our APIs. This uses the 'double submit' method requiring both
a cookie and custom header on mutating requests.

A CSRF_SECURE_COOKIE environment variable is added to allow for
dev access over insecure connections. Default is "true" which sets
the Secure attribute on the anti-CSRF token cookie, meaning browsers
will not send this cookie over an insecure connection. We override
this in the dev overlays to set it to "false".